refactor: 기존 TapGameView 대치
